[OOP_INITIAL] Created Multiple Header and Sorted Out Libraries for Future Usage
Status: All files is not yet completed.
Remember: We have to understand how header files works.

NOTE: Here in this commit, we did seperate everything so that everything is sorted out or able to be fixed by file. This will be the most risky attempt because if we did this on the long run, chances are we might able to identify or how to do this anymore. It is a good choice but dangerous. However, go forth!