[BASE WORKFLOW] Added Essential Files for System Creation
NOTE: At this point, we are about to give this whole commit for the last time. On the next commit we are gonna be commiting files individually due to the fact that we have to keep tracking changes on single files, not on multiple files. This will be a rule of thumb and we are trying to be practical here. For now, no futher classic information regarding of what I did.

Keep in mind that it may work. or not. We will find out soon.