[HUGE DEVELOPMENT] -> Unrecorded Changes, 09/15/2019
NOTE -> Due to time constraints, I cannot track changes as the week goes down. I have to do fast development than before. But keep in mind that things gets broken due to structure change of storage for each object that we have. Meaning that instead of storing time or date in std::string. I have to store them in struct tm. Which in the end clearly shows that time structure in database is not the datastamp style but rather diced for now. This is just an initial statement and I'm still not sure if that kind of approach would lead us in the right moment. But anyway, ever since we finished doing this exam. I gotta go fast.